requirements.txt
streamlit>=1.37
torch>=2.2
torchvision>=0.17
timm>=0.9.2
pillow>=10.0
numpy>=1.26
matplotlib>=3.8